init: java/jni project (#76)
# Goal
The goal of this PR is to setup the folders and project for java and jni

Closes #63 

# Discussion
- Currently we are not using `jni/cbindgen.toml` directly. The reason is
that we are exposing struct and types and enums in our api that is not
easily translatable to java. The way Signal works is that they only
expose functions and they have a python script that uses `bindgen`
output and generates a `Native.java`. In that way you can only change
the `Rust` side and it would generate the compatible the native api
signature in `java`. The catch is that it would be harder to expose
`Structs` and `enums` and non primitive types in the api and you would
need to serialize and deserialize to move everything between `java` and
`rust` that is not a primitive type. I'm looking into `JNI` more deeply
to understand what are the best practices and if it does make sense for
us to do something like that or not.
- Currently the example `hello` jni is being called from java and
executed in rust so the connection is established.
- gradle related files are all auto-generated, we might need to modify
them in future but it looks fine for now

# Checklist
- [x] Tests added for jni connection